More refactoring/cleanup
