+++ main.js
@@ -34,14 +34,7 @@ if (cluster.isMaster) {
   console.log("Master setup complete");
   
 } else {
-  // Workers can share any TCP connection
-  // In this case it is an HTTP server
-  // http.createServer((req, res) => {
-  //   res.writeHead(200);
-  //   res.end('hello world\n');
-  // }).listen(8000);
 
-  
   console.log("CHILD", cluster.worker.id);
   
   MainAlgorithms.ComputeSingleSliceValidSolutions(cluster.worker.id-1, numCPUs);
